Changed primary list ADT to array-based
After considering the amount of direct indexing necessary in the
required methods, changed to an array-based list. This may have a
tradeoff in efficiency when it comes to populating the list and sorting
it initially, but when accessing elements to be printed will be much
more efficient. Consider using quicksort instead of mergesort.